Fix various signed/unsigned mismatches.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/server/trunk@15197 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ common/holy.c
@@ -146,7 +146,7 @@ const object *pntr_to_god_obj(godlink *godlnk) {
  * @param maxlen maximum wanted length of the description, if 0 no maximum length.
  * @return information actually written, combination of @ref GOD_xxx "GOD_xxx", based on maxlen.
  */
-int describe_god(const object *god, int what, StringBuffer *buf, int maxlen) {
+int describe_god(const object *god, int what, StringBuffer *buf, size_t maxlen) {
     int hassomething = 0, real = 0;
     char temp[HUGE_BUF];
     StringBuffer *add;
